+++ examples/set_ops.rs
@@ -1,4 +1,5 @@
 //! Demonstrates set operations with `SmolBitmap`
+#![allow(clippy::many_single_char_names)]
 
 use smol_bitmap::SmolBitmap;